extracted BezierDrawer that allows drawing a bezier curve with specified texture from three points.
